feat: easier install (#36)
* feat: change the interface to projection layer.

This commit switches the projection layer from
flax nn module to general class.

* feat: generate and solve simple QP benchmark

This commit:
- Adds a script to generate a simple QP
benchmark. The parameter affects the equality
constraint.
- Adds a script to setup, train, and test
an HCNN on the simple QP problem.

* feat: Run simple QP benchmarks

This commit:
- Introduces loader and updates the generation of
the benchmarks.
- Adds some of the datasets for convenience.
- Updates the interface of the project class to
account for changes in some of the constraint parameters.
- Update the equality constraint project method for
the cases of varying parameters.
- Generated admm iteration with appropriate signatures
for the different varying parameters.
- Added test for the project class.

* feat: enable installation as package

* chore: update dependencies

---------

Co-authored-by: pgrontas <pgrontas@control.ee.ethz.ch>